--- scripts/bundles/clean-css.js
@@ -1,1 +1,0 @@
-exports.cleanCss = require('clean-css');
